Pagos + Updates + Fix
|+ Pagos:
|- CRUD Pagos
|
|+ Updates:
|- Se movio Agendamientos del menu lateral a la barra superior
|
|+ Fix:
|- Create Cotizacion el boton de submit era habilitado al hacer check en "Todos" sin importar si hubiran items con checked en true o no

+++ app/Http/Controllers/Admin/PreevaluacionController.php
@@ -198,6 +198,7 @@ class PreevaluacionController extends Controller
      */
     public function destroy(Preevaluacion $preevaluacion)
     {
+      $this->authorize('destroy', [Preevaluacion::class, $proceso]);
       if($preevaluacion->delete()){
         return redirect()->route('admin.proceso.show', ['proceso' => $preevaluacion->proceso_id])->with([
                 'flash_class'   => 'alert-success',
